Refactor PreparedPut tests
