+++ storio-common/src/main/java/com/pushtorefresh/storio/operation/PreparedOperation.java
@@ -1,7 +1,6 @@
 package com.pushtorefresh.storio.operation;
 
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import rx.Observable;
 
@@ -17,7 +16,7 @@ public interface PreparedOperation<Result> {
      *
      * @return result of operation
      */
-    @Nullable
+    @NonNull
     Result executeAsBlocking();
 
     /**
